docker: Add initial Docker/Containerfile
Having a local Containerfile allows anybody to build rsyslog containers.
Better yet, it allows rsyslog to supply official Docker containers for
users to use, rather then 'random containers from the internet'.

Note, that Containerfile is a modern way to signify agnostic Containers.
Internally, they are 100% pure Dockerfile's but for wider acceptance and
inclusion, Containerfile is the modern take.

The `--file Containerfile` argument might be needed for whichever
builder is used.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>